Refactor CreateChallengeForm and extract validation logic into separate hook and utility functions
